Refactor again
I need the form fields hashing logic to be available in both
support-frontend and support-workers.

Also removed an unused import.
